Module reqwest::async

An 'async' implementation of the reqwest Client.

Relies on the futures crate, which is unstable, hence this module is unstable.

Structs

Body
An asynchronous Stream.

Chunk
A chunk of bytes for a Body.

Client
An asynchronous Client to make Requests with.

ClientBuilder
A ClientBuilder can be used to create a Client with custom configuration:

Decoder
A response decompressor over a non-blocking stream of chunks.

Request
A request which can be executed with Client::execute().

RequestBuilder
A builder to construct the properties of a Request.

Response
A Response to a submitted Request.