Trait reqwest::async::ResponseBuilderExt

pub trait ResponseBuilderExt {
    fn url(&mut self, url: Url) -> &mut Self;
}

Extension trait for http::response::Builder objects

Allows the user to add a Url to the http::Response

Required methods

fn url(&mut self, url: Url) -> &mut Self

A builder method for the http::response::Builder type that allows the user to add a Url to the http::Response

Example

use url::Url;
use http::response::Builder;
use reqwest::async::ResponseBuilderExt;
let response = Builder::new()
    .status(200)
    .url(Url::parse("http://example.com")?)
    .body(())?;

Loading content...