reqwest::blocking

Function get

pub fn get<T: IntoUrl>(url: T) -> Result<Response>

Shortcut method to quickly make a blocking GET request.

NOTE: This function creates a new internal Client on each call, and so should not be used if making many requests. Create a Client instead.

Examples

let body = reqwest::blocking::get("https://www.rust-lang.org")?
    .text()?;

Errors

This function fails if:

the native TLS backend cannot be initialized,

the supplied Url cannot be parsed,

there was an error while sending request,

a redirect loop was detected,

the redirect limit was exhausted, or

the total download time exceeds 30 seconds.