reqwest::cookie
Struct Jar

pub struct Jar(/* private fields */);

A good default CookieStore implementation.

This is the implementation used when simply calling cookie_store(true). This type is exposed to allow creating one and filling it with some existing cookies more easily, before creating a Client.

For more advanced scenarios, such as needing to serialize the store or manipulate it between requests, you may refer to the reqwest_cookie_store crate.

Implementations

impl Jar

pub fn add_cookie_str(&self, cookie: &str, url: &Url)

Add a cookie to this jar.

Example

use reqwest::{cookie::Jar, Url};

let cookie = "foo=bar; Domain=yolo.local";
let url = "https://yolo.local".parse::<Url>().unwrap();

let jar = Jar::default();
jar.add_cookie_str(cookie, &url);

// and now add to a `ClientBuilder`?