Struct reqwest::header::GetAll

pub struct GetAll<'a, T> { /* private fields */ }

A view to all values stored in a single entry.

This struct is returned by HeaderMap::get_all.

Implementations

impl<'a, T> GetAll<'a, T> where
T: 'a,

pub fn iter(&self) -> ValueIter<'a, T>

Returns an iterator visiting all values associated with the entry.

Values are iterated in insertion order.

Examples

let mut map = HeaderMap::new();
map.insert(HOST, "hello.world".parse().unwrap());
map.append(HOST, "hello.earth".parse().unwrap());

let values = map.get_all("host");
let mut iter = values.iter();
assert_eq!(&"hello.world", iter.next().unwrap());
assert_eq!(&"hello.earth", iter.next().unwrap());
assert!(iter.next().is_none());