Struct reqwest::header::LastEventId

pub struct LastEventId(pub String);

Last-Event-ID header, defined in RFC3864

The Last-Event-ID header contains information about the last event in an http interaction so that it's easier to track of event state. This is helpful when working with Server-Sent-Events. If the connection were to be dropped, for example, it'd be useful to let the server know what the last event you received was.

The spec is a String with the id of the last event, it can be an empty string which acts a sort of "reset".

Example

use hyper::header::{Headers, LastEventId};

let mut headers = Headers::new();
headers.set(LastEventId("1".to_owned()));