reqwest::multipart
Struct Part

pub struct Part { /* private fields */ }

A field in a multipart form.

Implementations

impl Part

pub fn text<T>(value: T) -> Part
where T: Into<Cow<'static, str>>,

Makes a text parameter.

pub fn bytes<T>(value: T) -> Part
where T: Into<Cow<'static, [u8]>>,

Makes a new parameter from arbitrary bytes.

pub fn stream<T: Into<Body>>(value: T) -> Part

Makes a new parameter from an arbitrary stream.

pub fn stream_with_length<T: Into<Body>>(value: T, length: u64) -> Part

Makes a new parameter from an arbitrary stream with a known length. This is particularly useful when adding something like file contents as a stream, where you can know the content length beforehand.

pub async fn file<T: AsRef<Path>>(path: T) -> Result<Part>

Makes a file parameter.

Errors

Errors when the file cannot be opened.

pub fn mime_str(self, mime: &str) -> Result<Part>

Tries to set the mime of this part.

pub fn file_name<T>(self, filename: T) -> Part
where T: Into<Cow<'static, str>>,

Sets the filename, builder style.

pub fn headers(self, headers: HeaderMap) -> Part

Sets custom headers for the part.