Module reqwest::redirect

Redirect Handling

By default, a Client will automatically handle HTTP redirects, having a maximum redirect chain of 10 hops. To customize this behavior, a redirect::Policy can be used with a ClientBuilder.

Structs

Action

An action to perform when a redirect status code is found.

Attempt

A type that holds information on the next request and previous requests in redirect chain.

Policy

A type that controls the policy on how to handle the following of redirects.