Struct reqwest::redirect::Policy

pub struct Policy { /* fields omitted */ }

A type that controls the policy on how to handle the following of redirects.

The default value will catch redirect loops, and has a maximum of 10 redirects it will follow in a chain before returning an error.

limited can be used have the same as the default behavior, but adjust the allowed maximum redirect hops in a chain.

none can be used to disable all redirect behavior.

custom can be used to create a customized policy.

Methods

impl Policy

pub fn limited(max: usize) -> Self

Create a Policy with a maximum number of redirects.

An Error will be returned if the max is reached.

pub fn none() -> Self

Create a Policy that does not follow any redirect.

pub fn custom<T>(policy: T) -> Self where
T: Fn(Attempt) -> Action + Send + Sync + 'static,

Create a custom Policy using the passed function.

Note

The default Policy handles a maximum loop chain, but the custom variant does not do that for you automatically. The custom policy should have some way of handling those.

Information on the next request and previous requests can be found on the Attempt argument passed to the closure.

Actions can be conveniently created from methods on the Attempt.

Example

let custom = redirect::Policy::custom(|attempt| {
    if attempt.previous().len() > 5 {
        attempt.error("too many redirects")
    } else if attempt.url().host_str() == Some("example.domain") {
        // prevent redirects to 'example.domain'
        attempt.stop()
    } else {
        attempt.follow()
    }
});
let client = reqwest::Client::builder()
    .redirect(custom)
    .build()?;

pub fn redirect(&self, attempt: Attempt) -> Action

Apply this policy to a given [Attempt] to produce a [Action].

Note

This method can be used together with Policy::custom() to construct one Policy that wraps another.

Example

let custom = redirect::Policy::custom(|attempt| {
    eprintln!("{}, Location: {:?}", attempt.status(), attempt.url());
    redirect::Policy::default().redirect(attempt)
});