Struct reqwest::RedirectAttempt

pub struct RedirectAttempt<'a> { /* fields omitted */ }

A type that holds information on the next request and previous requests in redirect chain.

Methods

impl<'a> RedirectAttempt<'a>

pub fn status(&self) -> StatusCode

Get the type of redirect.

pub fn url(&self) -> &Url

Get the next URL to redirect to.

Important traits for &'a mut [u8]
impl<'a> Write for &'a mut [u8]impl<'a> Read for &'a [u8]

pub fn previous(&self) -> &[Url]

Get the list of previous URLs that have already been requested in this chain.

pub fn follow(self) -> RedirectAction

Returns an action meaning reqwest should follow the next URL.

pub fn stop(self) -> RedirectAction

Returns an action meaning reqwest should not follow the next URL.

The 30x response will be returned as the Ok result.

pub fn loop_detected(self) -> RedirectAction

Returns an action meaning there was a loop of redirects found.

An Error will be returned for the result of the sent request.

pub fn too_many_redirects(self) -> RedirectAction

Returns an action meaning there was a loop of redirects found.

An Error will be returned for the result of the sent request.