Struct reqwest::RedirectPolicy

pub struct RedirectPolicy { /* fields omitted */ }

A type that controls the policy on how to handle the following of redirects.

The default value will catch redirect loops, and has a maximum of 10 redirects it will follow in a chain before returning an error.

Methods

impl RedirectPolicy

pub fn limited(max: usize) -> RedirectPolicy

Create a RedirectPolicy with a maximum number of redirects.

An Error will be returned if the max is reached.

pub fn none() -> RedirectPolicy

Create a RedirectPolicy that does not follow any redirect.

pub fn custom<T>(policy: T) -> RedirectPolicy where
T: Fn(RedirectAttempt) -> RedirectAction + Send + Sync + 'static,

Create a custom RedirectPolicy using the passed function.

Note

The default RedirectPolicy handles redirect loops and a maximum loop chain, but the custom variant does not do that for you automatically. The custom policy should have some way of handling those.

Information on the next request and previous requests can be found on the RedirectAttempt argument passed to the closure.

Actions can be conveniently created from methods on the RedirectAttempt.

Example

let custom = RedirectPolicy::custom(|attempt| {
    if attempt.previous().len() > 5 {
        attempt.too_many_redirects()
    } else if attempt.url().host_str() == Some("example.domain") {
        // prevent redirects to 'example.domain'
        attempt.stop()
    } else {
        attempt.follow()
    }
});
let client = reqwest::Client::builder()?
    .redirect(custom)
    .build()?;