Trait reqwest::cookie::CookieStore

pub trait CookieStore: Send + Sync {
    fn set_cookies(
        &self, 
        cookie_headers: &mut dyn Iterator<Item = &HeaderValue>, 
        url: &Url
    );
    fn cookies(&self, url: &Url) -> Option<HeaderValue>;
}

Actions for a persistent cookie store providing session support.

Required Methods

fn set_cookies(
&self,
cookie_headers: &mut dyn Iterator<Item = &HeaderValue>,
url: &Url
)

Store a set of Set-Cookie header values received from url

fn cookies(&self, url: &Url) -> Option<HeaderValue>

Get any Cookie values in the store for url