reqwest

Module dns

Available on non-WebAssembly only.

DNS resolution

Structs

Name

A name that must be resolved to addresses.

Traits

IntoResolvedocsrs

A more general trait implemented for types implementing Resolve.

Resolve

Trait for customizing DNS resolution in reqwest.

Type Aliases

Addrs

Alias for an Iterator trait object over SocketAddr.

Resolving

Alias for the Future type returned by a DNS resolver.