Struct reqwest::Body

pub struct Body { /* fields omitted */ }

Body type for a request.

Methods

impl Body

fn new<R: Read + Send + 'static>(reader: R) -> Body

Instantiate a Body from a reader.

Note

While allowing for many types to be used, these bodies do not have a way to reset to the beginning and be reused. This means that when encountering a 307 or 308 status code, instead of repeating the request at the new location, the Response will be returned with the redirect status code set.

A Body constructed from a set of bytes, like String or Vec<u8>, are stored differently and can be reused.