Struct reqwest::Body

pub struct Body { /* private fields */ }

An asynchronous request body.

Implementations

impl Body

pub fn as_bytes(&self) -> Option<&[u8]>

Returns a reference to the internal data of the Body.

None is returned, if the underlying data is a stream.