Struct reqwest::Client

pub struct Client { /* fields omitted */ }

An asynchronous Client to make Requests with.

The Client has various configuration values to tweak, but the defaults are set to what is usually the most commonly desired value. To configure a Client, use Client::builder().

The Client holds a connection pool internally, so it is advised that you create one and reuse it.

Methods

impl Client

pub fn new() -> Client

Constructs a new Client.

Panics

This method panics if TLS backend cannot initialized, or the resolver cannot load the system configuration.

Use Client::builder() if you wish to handle the failure as an Error instead of panicking.

pub fn builder() -> ClientBuilder

Creates a ClientBuilder to configure a Client.

This is the same as ClientBuilder::new().

pub fn get<U: IntoUrl>(&self, url: U) -> RequestBuilder

Convenience method to make a GET request to a URL.

Errors

This method fails whenever supplied Url cannot be parsed.

pub fn post<U: IntoUrl>(&self, url: U) -> RequestBuilder

Convenience method to make a POST request to a URL.

Errors

This method fails whenever supplied Url cannot be parsed.

pub fn put<U: IntoUrl>(&self, url: U) -> RequestBuilder

Convenience method to make a PUT request to a URL.

Errors

This method fails whenever supplied Url cannot be parsed.

pub fn patch<U: IntoUrl>(&self, url: U) -> RequestBuilder

Convenience method to make a PATCH request to a URL.

Errors

This method fails whenever supplied Url cannot be parsed.

pub fn delete<U: IntoUrl>(&self, url: U) -> RequestBuilder

Convenience method to make a DELETE request to a URL.

Errors

This method fails whenever supplied Url cannot be parsed.

pub fn head<U: IntoUrl>(&self, url: U) -> RequestBuilder

Convenience method to make a HEAD request to a URL.

Errors

This method fails whenever supplied Url cannot be parsed.

pub fn request<U: IntoUrl>(&self, method: Method, url: U) -> RequestBuilder

Start building a Request with the Method and Url.

Returns a RequestBuilder, which will allow setting headers and request body before sending.

Errors

This method fails whenever supplied Url cannot be parsed.

pub fn execute(
&self,
request: Request
) -> impl Future<Output = Result<Response, Error>>

Executes a Request.

A Request can be built manually with Request::new() or obtained from a RequestBuilder with RequestBuilder::build().

You should prefer to use the RequestBuilder and RequestBuilder::send().

Errors

This method fails if there was an error while sending request, redirect loop was detected or redirect limit was exhausted.