Struct reqwest::Proxy

pub struct Proxy { /* fields omitted */ }

Configuration of a proxy that a Client should pass requests to.

A Proxy has a couple pieces to it:

a URL of how to talk to the proxy

rules on what Client requests should be directed to the proxy

For instance, let's look at Proxy::http:

let proxy = reqwest::Proxy::http("https://secure.example")?;

This proxy will intercept all HTTP requests, and make use of the proxy at https://secure.example. A request to http://hyper.rs will talk to your proxy. A request to https://hyper.rs will not.

Multiple Proxy rules can be configured for a Client. The Client will check each Proxy in the order it was added. This could mean that a Proxy added first with eager intercept rules, such as Proxy::all, would prevent a Proxy later in the list from ever working, so take care.

Methods

impl Proxy

pub fn http<U: IntoUrl>(url: U) -> Result<Proxy>

Proxy all HTTP traffic to the passed URL.

Example

let client = reqwest::Client::builder()
    .proxy(reqwest::Proxy::http("https://my.prox")?)
    .build()?;

pub fn https<U: IntoUrl>(url: U) -> Result<Proxy>

Proxy all HTTPS traffic to the passed URL.

Example

let client = reqwest::Client::builder()
    .proxy(reqwest::Proxy::https("https://example.prox:4545")?)
    .build()?;

pub fn all<U: IntoUrl>(url: U) -> Result<Proxy>

Proxy all traffic to the passed URL.

Example

let client = reqwest::Client::builder()
    .proxy(reqwest::Proxy::all("http://pro.xy")?)
    .build()?;

pub fn custom<F>(fun: F) -> Proxy where
F: Fn(&Url) -> Option<Url> + Send + Sync + 'static,

Provide a custom function to determine what traffix to proxy to where.

Example

let target = reqwest::Url::parse("https://my.prox")?;
let client = reqwest::Client::builder()
    .proxy(reqwest::Proxy::custom(move |url| {
        if url.host_str() == Some("hyper.rs") {
            Some(target.clone())
        } else {
            None
        }
    }))
    .build()?;