reqwest

Struct RequestBuilder

pub struct RequestBuilder { /* private fields */ }

A builder to construct the properties of a Request.

To construct a RequestBuilder, refer to the Client documentation.

Implementations

impl RequestBuilder

pub fn from_parts(client: Client, request: Request) -> RequestBuilder

Assemble a builder starting from an existing Client and a Request.

pub fn header<K, V>(self, key: K, value: V) -> RequestBuilder
where HeaderName: TryFrom<K>, <HeaderName as TryFrom<K>>::Error: Into<Error>, HeaderValue: TryFrom<V>, <HeaderValue as TryFrom<V>>::Error: Into<Error>,

Add a Header to this Request.

pub fn headers(self, headers: HeaderMap) -> RequestBuilder

Add a set of Headers to the existing ones on this Request.

The headers will be merged in to any already set.

pub fn basic_auth<U, P>( self, username: U, password: Option<P>, ) -> RequestBuilder
where U: Display, P: Display,

Enable HTTP basic authentication.

let client = reqwest::Client::new();
let resp = client.delete("http://httpbin.org/delete")
    .basic_auth("admin", Some("good password"))
    .send()
    .await?;

pub fn bearer_auth<T>(self, token: T) -> RequestBuilder
where T: Display,

Enable HTTP bearer authentication.

pub fn body<T: Into<Body>>(self, body: T) -> RequestBuilder

Set the request body.

pub fn timeout(self, timeout: Duration) -> RequestBuilder

Enables a request timeout.

The timeout is applied from when the request starts connecting until the response body has finished. It affects only this request and overrides the timeout configured using ClientBuilder::timeout().

pub fn query<T: Serialize + ?Sized>(self, query: &T) -> RequestBuilder

Modify the query string of the URL.

Modifies the URL of this request, adding the parameters provided. This method appends and does not overwrite. This means that it can be called multiple times and that existing query parameters are not overwritten if the same key is used. The key will simply show up twice in the query string. Calling .query(&[("foo", "a"), ("foo", "b")]) gives "foo=a&foo=b".

Note

This method does not support serializing a single key-value pair. Instead of using .query(("key", "val")), use a sequence, such as .query(&[("key", "val")]). It’s also possible to serialize structs and maps into a key-value pair.

Errors

This method will fail if the object you provide cannot be serialized into a query string.

pub fn version(self, version: Version) -> RequestBuilder

Set HTTP version

pub fn form<T: Serialize + ?Sized>(self, form: &T) -> RequestBuilder

Send a form body.

Sets the body to the url encoded serialization of the passed value, and also sets the Content-Type: application/x-www-form-urlencoded header.

let mut params = HashMap::new();
params.insert("lang", "rust");

let client = reqwest::Client::new();
let res = client.post("http://httpbin.org")
    .form(&params)
    .send()
    .await?;

Errors

This method fails if the passed value cannot be serialized into url encoded format

pub fn json<T: Serialize + ?Sized>(self, json: &T) -> RequestBuilder

Send a JSON body.

Optional

This requires the optional json feature enabled.

Errors

Serialization can fail if T’s implementation of Serialize decides to fail, or if T contains a map with non-string keys.

pub fn build(self) -> Result<Request>

Build a Request, which can be inspected, modified and executed with Client::execute().

pub fn build_split(self) -> (Client, Result<Request>)

Build a Request, which can be inspected, modified and executed with Client::execute().

This is similar to RequestBuilder::build(), but also returns the embedded Client.

pub fn send(self) -> impl Future<Output = Result<Response, Error>>

Constructs the Request and sends it to the target URL, returning a future Response.

Errors

This method fails if there was an error while sending request, redirect loop was detected or redirect limit was exhausted.

Example

let response = reqwest::Client::new()
    .get("https://hyper.rs")
    .send()
    .await?;

pub fn try_clone(&self) -> Option<RequestBuilder>

Attempt to clone the RequestBuilder.

None is returned if the RequestBuilder can not be cloned, i.e. if the request body is a stream.

Examples

let client = reqwest::Client::new();
let builder = client.post("http://httpbin.org/post")
    .body("from a &str!");
let clone = builder.try_clone();
assert!(clone.is_some());