reqwest

Struct Response

pub struct Response { /* private fields */ }

Available on WebAssembly only.

A Response to a submitted Request.

Implementations

impl Response

pub fn status(&self) -> StatusCode

Get the StatusCode of this Response.

pub fn headers(&self) -> &HeaderMap

Get the Headers of this Response.

pub fn headers_mut(&mut self) -> &mut HeaderMap

Get a mutable reference to the Headers of this Response.

pub fn content_length(&self) -> Option<u64>

Get the content-length of this response, if known.

Reasons it may not be known:

The server didn’t send a content-length header.

The response is compressed and automatically decoded (thus changing the actual decoded length).

pub fn url(&self) -> &Url

Get the final Url of this Response.

pub async fn json<T: DeserializeOwned>(self) -> Result<T>

Available on crate feature json only.

Try to deserialize the response body as JSON.

pub async fn text(self) -> Result<String>

Get the response text.

pub async fn bytes(self) -> Result<Bytes>

Get the response as bytes

pub fn bytes_stream(self) -> impl Stream<Item = Result<Bytes>>

Available on crate feature stream only.

Convert the response into a Stream of Bytes from the body.

pub fn error_for_status(self) -> Result<Self>

Turn a response into an error if the server returned an error.

pub fn error_for_status_ref(&self) -> Result<&Self>

Turn a reference to a response into an error if the server returned an error.