Struct reqwest::Response

pub struct Response { /* fields omitted */ }

A Response to a submitted Request.

Implementations

impl Response

pub fn status(&self) -> StatusCode

Get the StatusCode of this Response.

Examples

Checking for general status class:

let resp = reqwest::get("http://httpbin.org/get")?;
if resp.status().is_success() {
    println!("success!");
} else if resp.status().is_server_error() {
    println!("server error!");
} else {
    println!("Something else happened. Status: {:?}", resp.status());
}

Checking for specific status codes:

use reqwest::Client;
use reqwest::StatusCode;
let client = Client::new();

let resp = client.post("http://httpbin.org/post")
    .body("possibly too large")
    .send()?;

match resp.status() {
    StatusCode::OK => println!("success!"),
    StatusCode::PAYLOAD_TOO_LARGE => {
        println!("Request payload is too large!");
    }
    s => println!("Received response status: {:?}", s),
};

pub fn headers(&self) -> &HeaderMap

Get the Headers of this Response.

Example

Saving an etag when caching a file:

use reqwest::Client;
use reqwest::header::ETAG;

let client = Client::new();

let mut resp = client.get("http://httpbin.org/cache").send()?;
if resp.status().is_success() {
    if let Some(etag) = resp.headers().get(ETAG) {
        std::fs::write("etag", etag.as_bytes());
    }
    let mut file = std::fs::File::create("file")?;
    resp.copy_to(&mut file)?;
}

pub fn cookies<'a>(&'a self) -> impl Iterator<Item = Cookie<'a>> + 'a

Retrieve the cookies contained in the response.

Note that invalid ‘Set-Cookie’ headers will be ignored.

pub fn version(&self) -> Version

Get the HTTP Version of this Response.

pub fn url(&self) -> &Url

Get the final Url of this Response.

Example

let resp = reqwest::get("http://httpbin.org/redirect/1")?;
assert_eq!(resp.url().as_str(), "http://httpbin.org/get");

pub fn remote_addr(&self) -> Option<SocketAddr>

Get the remote address used to get this Response.

Example

let resp = reqwest::get("http://httpbin.org/redirect/1")?;
println!("httpbin.org address: {:?}", resp.remote_addr());

pub fn content_length(&self) -> Option<u64>

Get the content-length of the response, if it is known.

Reasons it may not be known:

The server didn’t send a content-length header.

The response is gzipped and automatically decoded (thus changing the actual decoded length).

pub fn json<T: DeserializeOwned>(&mut self) -> Result<T>

Try and deserialize the response body as JSON using serde.

Examples

#[derive(Deserialize)]
struct Ip {
    origin: String,
}

let json: Ip = reqwest::get("http://httpbin.org/ip")?.json()?;

Errors

This method fails whenever the response body is not in JSON format or it cannot be properly deserialized to target type T. For more details please see serde_json::from_reader. serde_json::from_reader: https://docs.serde.rs/serde_json/fn.from_reader.html

pub fn text(&mut self) -> Result<String>

Get the response text.

This method decodes the response body with BOM sniffing and with malformed sequences replaced with the REPLACEMENT CHARACTER. Encoding is determinated from the charset parameter of Content-Type header, and defaults to utf-8 if not presented.

Example

let content = reqwest::get("http://httpbin.org/range/26")?.text()?;

Note

This consumes the body. Trying to read more, or use of response.json() will return empty values.

pub fn text_with_charset(&mut self, default_encoding: &str) -> Result<String>

Get the response text given a specific encoding.

This method decodes the response body with BOM sniffing and with malformed sequences replaced with the REPLACEMENT CHARACTER. You can provide a default encoding for decoding the raw message, while the charset parameter of Content-Type header is still prioritized. For more information about the possible encoding name, please go to https://docs.rs/encoding_rs/0.8.17/encoding_rs/#relationship-with-windows-code-pages

Example

let content = reqwest::get("http://httpbin.org/range/26")?.text_with_charset("utf-8")?;

Note

This consumes the body. Trying to read more, or use of response.json() will return empty values.

pub fn copy_to<W: ?Sized>(&mut self, w: &mut W) -> Result<u64> where
W: Write,

Copy the response body into a writer.

This function internally uses std::io::copy and hence will continuously read data from the body and then write it into writer in a streaming fashion until EOF is met.

On success, the total number of bytes that were copied to writer is returned.

Example

let mut resp = reqwest::get("http://httpbin.org/range/5")?;
let mut buf: Vec<u8> = vec![];
resp.copy_to(&mut buf)?;
assert_eq!(b"abcde", buf.as_slice());

pub fn error_for_status(self) -> Result<Self>

Turn a response into an error if the server returned an error.

Example

let res = reqwest::get("http://httpbin.org/status/400")?
    .error_for_status();
if let Err(err) = res {
    assert_eq!(err.status(), Some(reqwest::StatusCode::BAD_REQUEST));
}

pub fn error_for_status_ref(&self) -> Result<&Self>

Turn a reference to a response into an error if the server returned an error.

Example

let res = reqwest::get("http://httpbin.org/status/400")?;
let res = res.error_for_status_ref();
if let Err(err) = res {
    assert_eq!(err.status(), Some(reqwest::StatusCode::BAD_REQUEST));
}