Struct reqwest::Response

pub struct Response { /* private fields */ }

A Response to a submitted Request.

Implementations

impl Response

pub fn status(&self) -> StatusCode

Get the StatusCode of this Response.

pub fn version(&self) -> Version

Get the HTTP Version of this Response.

pub fn headers(&self) -> &HeaderMap

Get the Headers of this Response.

pub fn headers_mut(&mut self) -> &mut HeaderMap

Get a mutable reference to the Headers of this Response.

pub fn content_length(&self) -> Option<u64>

Get the content-length of this response, if known.

Reasons it may not be known:

The server didn’t send a content-length header.

The response is compressed and automatically decoded (thus changing the actual decoded length).

pub fn url(&self) -> &Url

Get the final Url of this Response.

pub fn remote_addr(&self) -> Option<SocketAddr>

Get the remote address used to get this Response.

pub async fn text(self) -> Result<String>

Get the full response text.

This method decodes the response body with BOM sniffing and with malformed sequences replaced with the REPLACEMENT CHARACTER. Encoding is determinated from the charset parameter of Content-Type header, and defaults to utf-8 if not presented.

Example

let content = reqwest::get("http://httpbin.org/range/26")
    .await?
    .text()
    .await?;

println!("text: {:?}", content);

pub async fn text_with_charset(self, default_encoding: &str) -> Result<String>

Get the full response text given a specific encoding.

This method decodes the response body with BOM sniffing and with malformed sequences replaced with the REPLACEMENT CHARACTER. You can provide a default encoding for decoding the raw message, while the charset parameter of Content-Type header is still prioritized. For more information about the possible encoding name, please go to encoding_rs docs.

Example

let content = reqwest::get("http://httpbin.org/range/26")
    .await?
    .text_with_charset("utf-8")
    .await?;

println!("text: {:?}", content);

pub async fn json<T: DeserializeOwned>(self) -> Result<T>

Try to deserialize the response body as JSON.

Optional

This requires the optional json feature enabled.

Examples

// This `derive` requires the `serde` dependency.
#[derive(Deserialize)]
struct Ip {
    origin: String,
}

let ip = reqwest::get("http://httpbin.org/ip")
    .await?
    .json::<Ip>()
    .await?;

println!("ip: {}", ip.origin);

Errors

This method fails whenever the response body is not in JSON format or it cannot be properly deserialized to target type T. For more details please see serde_json::from_reader.

pub async fn bytes(self) -> Result<Bytes>

Get the full response body as Bytes.

Example

let bytes = reqwest::get("http://httpbin.org/ip")
    .await?
    .bytes()
    .await?;

println!("bytes: {:?}", bytes);

pub async fn chunk(&mut self) -> Result<Option<Bytes>>

Stream a chunk of the response body.

When the response body has been exhausted, this will return None.

Example

let mut res = reqwest::get("https://hyper.rs").await?;

while let Some(chunk) = res.chunk().await? {
    println!("Chunk: {:?}", chunk);
}

pub fn error_for_status(self) -> Result<Self>

Turn a response into an error if the server returned an error.

Example

fn on_response(res: Response) {
    match res.error_for_status() {
        Ok(_res) => (),
        Err(err) => {
            // asserting a 400 as an example
            // it could be any status between 400...599
            assert_eq!(
                err.status(),
                Some(reqwest::StatusCode::BAD_REQUEST)
            );
        }
    }
}

pub fn error_for_status_ref(&self) -> Result<&Self>

Turn a reference to a response into an error if the server returned an error.

Example

fn on_response(res: &Response) {
    match res.error_for_status_ref() {
        Ok(_res) => (),
        Err(err) => {
            // asserting a 400 as an example
            // it could be any status between 400...599
            assert_eq!(
                err.status(),
                Some(reqwest::StatusCode::BAD_REQUEST)
            );
        }
    }
}