Module reqwest::tls

TLS configuration

By default, a Client will make use of system-native transport layer security to connect to HTTPS destinations. This means schannel on Windows, Security-Framework on macOS, and OpenSSL on Linux.

Additional X509 certificates can be configured on a ClientBuilder with the Certificate type.

Client certificates can be add to a ClientBuilder with the Identity type.

Various parts of TLS can also be configured or even disabled on the ClientBuilder.

Structs

Certificate

Represents a server X509 certificate.

Identity

Represents a private key and X509 cert as a client certificate.

Version

A TLS protocol version.