Trait reqwest::ResponseBuilderExt

pub trait ResponseBuilderExt {
    fn url(self, url: Url) -> Self;
}

Extension trait for http::response::Builder objects

Allows the user to add a Url to the http::Response

Required methods

fn url(self, url: Url) -> Self

A builder method for the http::response::Builder type that allows the user to add a Url to the http::Response