reqwest

Trait UnixSocketProvider

pub trait UnixSocketProvider { }

Available on non-WebAssembly and docsrs only.

A provider for Unix Domain Socket paths.

This trait is sealed. This allows us expand the support in the future by controlling who can implement the trait.

It’s available in the docs to see what type may be passed in.