Struct reqwest::unstable::async::ClientBuilder

pub struct ClientBuilder { /* fields omitted */ }

A ClientBuilder can be used to create a Client with custom configuration:

Methods

impl ClientBuilder

pub fn new() -> ClientBuilder

Constructs a new ClientBuilder

pub fn build(&mut self, handle: &Handle) -> Result<Client>

Returns a Client that uses this ClientBuilder configuration.

Errors

This method fails if native TLS backend cannot be initialized.

Panics

This method consumes the internal state of the builder. Trying to use this builder again after calling build will panic.

pub fn add_root_certificate(&mut self, cert: Certificate) -> &mut ClientBuilder

Add a custom root certificate.

This can be used to connect to a server that has a self-signed certificate for example.

pub fn identity(&mut self, identity: Identity) -> &mut ClientBuilder

Sets the identity to be used for client certificate authentication.

pub fn danger_disable_hostname_verification(&mut self) -> &mut ClientBuilder

Disable hostname verification.

Warning

You should think very carefully before you use this method. If hostname verification is not used, any valid certificate for any site will be trusted for use from any other. This introduces a significant vulnerability to man-in-the-middle attacks.

pub fn enable_hostname_verification(&mut self) -> &mut ClientBuilder

Enable hostname verification.

pub fn default_headers(&mut self, headers: Headers) -> &mut ClientBuilder

Sets the default headers for every request.

pub fn gzip(&mut self, enable: bool) -> &mut ClientBuilder

Enable auto gzip decompression by checking the ContentEncoding response header.

Default is enabled.

pub fn proxy(&mut self, proxy: Proxy) -> &mut ClientBuilder

Add a Proxy to the list of proxies the Client will use.

pub fn redirect(&mut self, policy: RedirectPolicy) -> &mut ClientBuilder

Set a RedirectPolicy for this client.

Default will follow redirects up to a maximum of 10.

pub fn referer(&mut self, enable: bool) -> &mut ClientBuilder

Enable or disable automatic setting of the Referer header.

Default is true.

pub fn timeout(&mut self, timeout: Duration) -> &mut ClientBuilder

Set a timeout for both the read and write operations of a client.

pub fn dns_threads(&mut self, threads: usize) -> &mut ClientBuilder

Set number of DNS threads.